Final
Still a bit glitchy with the AI paddle movement, but less than other attempts. Can't control the speed- my attempts at doing so negatively impacted the movement. Wondering if there's a better way to code this so that the movement is smoother? Also feel like my code isn't super efficient.